requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dlib==19.24.1
entrypoints==0.4
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.11.0
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.24.1
gradio_client==0.0.8
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
linkify-it-py==2.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.0.4
multidict==6.0.4
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.4
opencv-python==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.9
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.7
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.20.0
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tifffile==2023.3.21
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
